requirements.txt
streamlit
python-dotenv
requests
pydeck
plotly
